Use helper functions from integration tests package part 6 (WIP)
