+++ pkg/sdk/helper_test.go
@@ -3,7 +3,6 @@ package sdk
 import (
 	"context"
 	"testing"
-	"time"
 
 	"github.com/stretchr/testify/require"
 )
@@ -603,15 +602,15 @@ func cleanupDatabaseRoleProvider(t *testing.T, ctx context.Context, client *Clie
 //	return createAlertWithOptions(t, client, database, schema, warehouse, schedule, condition, action, &CreateAlertOptions{})
 //}
 
-func ParseTimestampWithOffset(s string) (*time.Time, error) {
-	t, err := time.Parse("2006-01-02T15:04:05-07:00", s)
-	if err != nil {
-		return nil, err
-	}
-	_, offset := t.Zone()
-	adjustedTime := t.Add(-time.Duration(offset) * time.Second)
-	return &adjustedTime, nil
-}
+//func ParseTimestampWithOffset(s string) (*time.Time, error) {
+//	t, err := time.Parse("2006-01-02T15:04:05-07:00", s)
+//	if err != nil {
+//		return nil, err
+//	}
+//	_, offset := t.Zone()
+//	adjustedTime := t.Add(-time.Duration(offset) * time.Second)
+//	return &adjustedTime, nil
+//}
 
 //func createUser(t *testing.T, client *Client) (*User, func()) {
 //	t.Helper()
@@ -708,57 +707,57 @@ func ParseTimestampWithOffset(s string) (*time.Time, error) {
 //	}, stageCleanup
 //}
 
-func createDynamicTable(t *testing.T, client *Client) (*DynamicTable, func()) {
-	t.Helper()
-	return createDynamicTableWithOptions(t, client, nil, nil, nil, nil)
-}
-
-func createDynamicTableWithOptions(t *testing.T, client *Client, warehouse *Warehouse, database *Database, schema *Schema, table *Table) (*DynamicTable, func()) {
-	t.Helper()
-	var warehouseCleanup func()
-	if warehouse == nil {
-		warehouse, warehouseCleanup = createWarehouse(t, client)
-	}
-	var databaseCleanup func()
-	if database == nil {
-		database, databaseCleanup = createDatabase(t, client)
-	}
-	var schemaCleanup func()
-	if schema == nil {
-		schema, schemaCleanup = createSchema(t, client, database)
-	}
-	var tableCleanup func()
-	if table == nil {
-		table, tableCleanup = createTable(t, client, database, schema)
-	}
-	name := NewSchemaObjectIdentifier(schema.DatabaseName, schema.Name, randomString(t))
-	targetLag := TargetLag{
-		Lagtime: String("2 minutes"),
-	}
-	query := "select id from " + table.ID().FullyQualifiedName()
-	comment := randomComment(t)
-	ctx := context.Background()
-	err := client.DynamicTables.Create(ctx, NewCreateDynamicTableRequest(name, warehouse.ID(), targetLag, query).WithOrReplace(true).WithComment(&comment))
-	require.NoError(t, err)
-	entities, err := client.DynamicTables.Show(ctx, NewShowDynamicTableRequest().WithLike(&Like{Pattern: String(name.Name())}).WithIn(&In{Schema: schema.ID()}))
-	require.NoError(t, err)
-	require.Equal(t, 1, len(entities))
-	return &entities[0], func() {
-		require.NoError(t, client.DynamicTables.Drop(ctx, NewDropDynamicTableRequest(name)))
-		if tableCleanup != nil {
-			tableCleanup()
-		}
-		if schemaCleanup != nil {
-			schemaCleanup()
-		}
-		if databaseCleanup != nil {
-			databaseCleanup()
-		}
-		if warehouseCleanup != nil {
-			warehouseCleanup()
-		}
-	}
-}
+//func createDynamicTable(t *testing.T, client *Client) (*DynamicTable, func()) {
+//	t.Helper()
+//	return createDynamicTableWithOptions(t, client, nil, nil, nil, nil)
+//}
+//
+//func createDynamicTableWithOptions(t *testing.T, client *Client, warehouse *Warehouse, database *Database, schema *Schema, table *Table) (*DynamicTable, func()) {
+//	t.Helper()
+//	var warehouseCleanup func()
+//	if warehouse == nil {
+//		warehouse, warehouseCleanup = createWarehouse(t, client)
+//	}
+//	var databaseCleanup func()
+//	if database == nil {
+//		database, databaseCleanup = createDatabase(t, client)
+//	}
+//	var schemaCleanup func()
+//	if schema == nil {
+//		schema, schemaCleanup = createSchema(t, client, database)
+//	}
+//	var tableCleanup func()
+//	if table == nil {
+//		table, tableCleanup = createTable(t, client, database, schema)
+//	}
+//	name := NewSchemaObjectIdentifier(schema.DatabaseName, schema.Name, randomString(t))
+//	targetLag := TargetLag{
+//		Lagtime: String("2 minutes"),
+//	}
+//	query := "select id from " + table.ID().FullyQualifiedName()
+//	comment := randomComment(t)
+//	ctx := context.Background()
+//	err := client.DynamicTables.Create(ctx, NewCreateDynamicTableRequest(name, warehouse.ID(), targetLag, query).WithOrReplace(true).WithComment(&comment))
+//	require.NoError(t, err)
+//	entities, err := client.DynamicTables.Show(ctx, NewShowDynamicTableRequest().WithLike(&Like{Pattern: String(name.Name())}).WithIn(&In{Schema: schema.ID()}))
+//	require.NoError(t, err)
+//	require.Equal(t, 1, len(entities))
+//	return &entities[0], func() {
+//		require.NoError(t, client.DynamicTables.Drop(ctx, NewDropDynamicTableRequest(name)))
+//		if tableCleanup != nil {
+//			tableCleanup()
+//		}
+//		if schemaCleanup != nil {
+//			schemaCleanup()
+//		}
+//		if databaseCleanup != nil {
+//			databaseCleanup()
+//		}
+//		if warehouseCleanup != nil {
+//			warehouseCleanup()
+//		}
+//	}
+//}
 
 //func createStageWithURL(t *testing.T, client *Client, name AccountObjectIdentifier, url string) (*Stage, func()) {
 //	t.Helper()
